requirements.txt
Flask
numpy
opencv-python
pytesseract
Pillow
scikit-learn
werkzeug
logging
